putting scripts and other code into modules for cleaner code
