+++ app.py
@@ -2,7 +2,9 @@ from flask import Flask, render_template, request, redirect, url_for, jsonify, s
 import json, os
 from flask_session import Session
 #importing scripts
-from scripts import get_program_codes, get_prereqs, prepare_data, utils
+from scripts.Getting_Info_For_Major import get_courses_of_major, get_info_for_major, get_prereqs
+from scripts import utils
+from scripts.app_functions import json_graph_file_handling, log_entry
 from datetime import datetime
 
 app = Flask(__name__)
@@ -17,115 +19,36 @@ courses_info = {}
 with open('static/json/courses_info.json', 'r', encoding='utf-8') as f:
     courses_info = json.load(f)
 
-@app.route('/', methods=['GET', "POST"])
+@app.route('/', methods=['GET', "POST"]) #home page
 def scrape_form():
-    
     #adding a log for each user query
-    user_ip = request.remote_addr
-    timestamp = datetime.now()
+    log_entry(request, "accessing home page")
 
-    if request.args.get("action") == "Scrape and Visualize":
-        print(f"user {user_ip} made a {request.args.get('action')} request for {request.args.get('url')} at {timestamp}")
-    else:
-        print(f"user {user_ip} made a {request.args.get('action')} request at {timestamp}")
-
-
-    if request.method == 'POST':
-
-        # when a user uploads a JSON file
-        action = request.form.get('action')
-        if action == "Load Graph":
-
-            if 'graphFile' not in request.files:
-                return render_template('scrape_form.html', error="No file selected for upload.")
+    action = request.args.get('action')
+    if request.method == 'POST' and action == "Load Graph": # when a user uploads a JSON file
+            if 'graphFile' not in request.files: #checking if it's a json file
+                return render_template('scrape_form.html', error="Invalid file json file provided")
             
             file = request.files['graphFile']
 
-            if file.filename == '':
-                return render_template('scrape_form.html', error="No file selected for upload.")
+            if not json_graph_file_handling(file): #check if valid json file
+                return render_template('scrape_form.html', error="Invalid file json file provided")
             
-            if file and file.filename.endswith('.json'):
-                try:
-                    loaded_data = json.load(file) # Parse the JSON file stream
-                    
-                    # Prepare data structures for index.html
-                    final_prereqs_for_template = {}
-                    final_details_for_template = {}
+            else: #process the json file for the graph
+                prereqs_data, details_data = json_graph_file_handling.process_json_graph_file(file)
 
-                    if 'nodes' not in loaded_data or 'edges' not in loaded_data:
-                        return render_template('scrape_form.html', error="Invalid JSON graph file: 'nodes' or 'edges' key missing.")
+                #Saving to session
+                session['prereqs_data'] = prereqs_data
+                session['details_data'] = details_data
+                session['graph_data_available'] = True
 
-                    # Process nodes from the loaded JSON
-                    for node_data in loaded_data.get('nodes', []):
-                        node_id = node_data.get('id')
-                        if not node_id: 
-                            print("Skipping node without ID from JSON:", node_data)
-                            continue 
+                return redirect("view_graph")
 
-                        # Extract details, using fallbacks for robustness
-                        sem_offered = node_data.get("original_semesters_offered", node_data.get("semesters_offered", "Unknown"))
-                        final_details_for_template[node_id] = {
-                            "title": node_data.get("original_title", node_data.get("title", "Unknown Title")),
-                            "credits": node_data.get("original_credits", node_data.get("credits", "N/A")),
-                            "semesters_offered": sem_offered,
-                            "color": node_data.get("color", utils.parse_semester_to_color(sem_offered)), # Use stored color or recalculate
-                            "id": node_id,
-                            "label": node_data.get("label", f"{node_id}\nUnknown Title\n(N/A credits)"),
-                            "shape": node_data.get("shape", "box"),
-                            "font": node_data.get("font", {'multi': 'html', 'align': 'center'}),
-                            "original_title": node_data.get("original_title", "Unknown Title"), # Ensure these are passed for editing
-                            "original_credits": node_data.get("original_credits", "N/A"),
-                            "original_semesters_offered": sem_offered,
-                            "x": node_data.get("x"),
-                            "y": node_data.get("y")
-                        }
-                        final_prereqs_for_template[node_id] = [] # Initialize prereqs list for this node
-
-                    # Process edges from the loaded JSON
-                    for edge_data in loaded_data.get('edges', []):
-                        from_node = edge_data.get('from')
-                        to_node = edge_data.get('to')
-                        if from_node and to_node:
-                            if to_node not in final_prereqs_for_template:
-                                final_prereqs_for_template[to_node] = []
-                                if to_node not in final_details_for_template: 
-                                     final_details_for_template[to_node] = {
-                                        "title": f"{to_node} (Prereq from Edge)", "credits": "N/A", 
-                                        "semesters_offered": "Unknown", "color": "grey", "id": to_node,
-                                        "label": f"{to_node}\n(Prereq from Edge)\n(N/A credits)", "shape": "box",
-                                        "font": {'multi': 'html', 'align': 'center'},
-                                        "original_title": f"{to_node} (Prereq from Edge)", "original_credits": "N/A",
-                                        "original_semesters_offered": "Unknown"
-                                     }
-                            final_prereqs_for_template[to_node].append(from_node)
-                    
-                    if not final_details_for_template: # Check if any nodes were actually processed
-                         return render_template('scrape_form.html', error="No valid node data found in the uploaded JSON file.")
-
-                    #Saving the variable to session
-                    session['prereqs_data'] = final_prereqs_for_template
-                    session['details_data'] = final_details_for_template
-                    session['graph_data_available'] = True
-
-                    return redirect("view_graph")
-                
-                except json.JSONDecodeError:
-                    return render_template('scrape_form.html', error="Invalid JSON file: Could not parse content.")
-                except Exception as e:
-                    print(f"Error processing uploaded graph file: {e}") # Log the actual error
-                    return render_template('scrape_form.html', error=f"An error occurred while processing the graph file: {e}")
-            else:
-                return render_template('scrape_form.html', error="Invalid file type. Please upload a .json file.")
-        else:
-            # If POST but not "Load Graph" action
-            return render_template('scrape_form.html', error="Unknown POST action.")
-
-    action = request.args.get('action')
     url = request.args.get('url')
     major = request.args.get("programSearch")
     
-    if action == "Visualize Program":
-        courses_prereqs_data, processed_details_data = prepare_data.process_program_data(url, major)
+    if action == "Visualize Program": 
+        courses_prereqs_data, processed_details_data = get_info_for_major.process_program_data(url, major) #scrape the major's site and grab all info
         
         #Saving the variable to session
         session['prereqs_data'] = courses_prereqs_data
@@ -141,51 +64,43 @@ def scrape_form():
             session.pop('prereqs_data', None)
             session.pop('details_data', None)
             session.pop('graph_data_available', None)
-            # This is an initial GET request to the form, no error needed
+            # This is an initial GET request to the form
             return render_template('scrape_form.html')
         else:
             session.pop('prereqs_data', None)
             session.pop('details_data', None)
             session.pop('graph_data_available', None)
-            # This could be a GET with an unknown action or some other case
             return render_template('scrape_form.html', error="Please select a valid action.")
 
-
-#see if there's a saved graph
-@app.route("/view_graph", methods=["GET","POST"])
-def view_graph():
-    if session.get('graph_data_available'):
+@app.route("/view_graph", methods=["GET","POST"]) #main route where graph is displayed
+def view_graph(): 
+    if session.get('graph_data_available'): #see if there's a saved graph
         prereqs = session.get('prereqs_data', {})
         details = session.get('details_data', {})
+        log_entry(request, "displaying graph")
         return render_template('index.html', prereqs=prereqs, details=details)
     else:
-        # If no data, redirect to home to load/scrape a program
+        # If no data, redirect back to home page
         return redirect(url_for('scrape_form'))
 
-@app.route("/add_program")
-def add_program():
-    if not session.get('graph_data_available'):
+@app.route("/add_program_form") # form for choosing which program to add to graph
+def add_program_form():
+    if not session.get('graph_data_available'): #verify there's an existing graph first
         return redirect(url_for('scrape_form', error="Please load or visualize a base program first."))
-    return render_template("add_network_form.html")
+    return render_template("add_program_form.html")
 
-@app.route("/add_program_to_graph", methods=["GET"])
+@app.route("/add_program_to_graph", methods=["GET"]) # adding another program to their graph (like a minor)
 def add_program_to_graph():
-    user_ip = request.remote_addr
-    timestamp = datetime.now()
-
     if not session.get('graph_data_available'):
         return redirect(url_for('scrape_form', error="No active graph to add to. Please start a new one."))
 
     new_program_url = request.args.get('url')
     new_program_name = request.args.get('programSearch') # From add_network_form.html
 
-    if not new_program_url or not new_program_name:
-        return render_template('add_network_form.html', error="Program URL or name missing.")
-
-    print(f"User {user_ip} attempting to add program {new_program_name} ({new_program_url}) at {timestamp}")
+    log_entry(request, f"adding program {new_program_name} (url: {new_program_url}) to graph")    
 
     # Fetch and process data for the new program
-    new_prereqs, new_details = prepare_data.process_program_data(new_program_url, new_program_name)
+    new_prereqs, new_details = get_info_for_major.process_program_data(new_program_url, new_program_name)
 
     if new_prereqs is None or new_details is None:
         return render_template('add_network_form.html', error=f"Could not process data for {new_program_name}.")
@@ -214,9 +129,8 @@ def add_program_to_graph():
     # Store merged data back in session
     session['prereqs_data'] = current_prereqs
     session['details_data'] = current_details
-    session['graph_data_available'] = True # Ensure this is set
+    session['graph_data_available'] = True 
 
-    print(f"Successfully merged {new_program_name} into the graph for user {user_ip} at {timestamp}")
     return redirect(url_for('view_graph'))
 
 @app.route('/modify_graph', methods=['GET'])
