+++ scripts/Getting_Info_For_Major/get_courses_of_major.py
@@ -1,7 +1,9 @@
 import requests
 from bs4 import BeautifulSoup
 
-#returns course codes of major
+#Given url of the major, it scrapes the site and gets all the courses of the major.
+#Eventually this could be ran on all programs and be hard-coded instead of running on every query
+
 def get_program_codes(url):
     res = requests.get(url)
     soup = BeautifulSoup(res.text, 'html.parser')
