+++ scripts/Scrape McGill's site (re-run once a year)/programs/get_programs_list.py
