+++ templates/index.html
@@ -39,7 +39,7 @@
       </header>
       <div class="custom-manipulation-buttons">
         <button id="exportGraphBtn">Export Graph (JSON)</button>
-        <a href="{{url_for('add_program')}}">
+        <a href="{{url_for('add_program_form')}}">
           <button id="addProgramToGraphBtn">
             Add Another Program to Graph
           </button>
